refactor(benchmarks): extract live viewport controllers
